Add FADE_IN and SLIDE effects and conditional ramp up and down
Ramp up and down of volume at start/end of note now only happens when the wave form changes.
Also, when the wave from remains the same, also ensured that there is no abrupt transition in the wave form. This is even the case when the volume or frequency changes.

Also added first version of SFX18. Still missing, a vibrato effect.